requirements.txt
altgraph==0.17.3
certifi==2022.12.7
charset-normalizer==2.0.12
click==7.1.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
easyocr==1.6.2
future==0.18.3
idna==3.4
imageio==2.15.0
importlib-metadata==4.8.3
itchat==1.2.32
kiwisolver==1.3.1
matplotlib==3.3.4
Naked==0.1.32
networkx==2.5.1
ninja==1.11.1
numpy==1.19.5
opencv-python-headless==4.5.4.60
packaging==21.3
pefile==2022.5.30
Pillow==8.4.0
prettytable==2.5.0
pyclipper==1.3.0.post4
pycryptodome==3.18.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt6==6.1.0
pyqt6-plugins==6.1.0.2.2
PyQt6-Qt6==6.2.3
PyQt6-sip==13.2.1
pyqt6-tools==6.1.0.3.2
PySimpleGUI==4.60.4
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
PyWavelets==1.1.1
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
qt6-applications==6.1.0.2.2
qt6-tools==6.1.0.1.2
requests==2.27.1
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.8.5.post1
shellescape==3.8.1
six==1.16.0
tifffile==2020.9.3
torch==1.10.2
torchvision==0.11.3
typing_extensions==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.6.0
win10toast==0.9